# size in log(n)