--- a/bench/Bench/Data/Map.purs
+++ b/bench/Bench/Data/Map.purs
@@ -1,0 +1,33 @@
+module Bench.Data.Map where
+
+import Prelude
+import Control.Monad.Eff (Eff)
+import Control.Monad.Eff.Console (CONSOLE, log)
+import Performance.Minibench (bench, benchWith)
+
+import Data.Tuple (Tuple(..))
+import Data.List as L
+import Data.Map as M
+
+benchMap :: Eff (console :: CONSOLE) Unit
+benchMap = benchSize
+  where
+  benchSize = do
+    let nats = L.range 0 999999
+        natPairs = (flip Tuple) unit <$> nats
+        singletonMap = M.singleton 0 unit
+        smallMap = M.fromFoldable $ L.take 100 natPairs
+        midMap = M.fromFoldable $ L.take 10000 natPairs
+        bigMap = M.fromFoldable $ natPairs
+
+    log "size: singleton map"
+    bench \_ -> M.size singletonMap
+
+    log $ "size: small map (" <> show (M.size smallMap) <> ")"
+    bench \_ -> M.size smallMap
+
+    log $ "size: midsize map (" <> show (M.size midMap) <> ")"
+    benchWith 100 \_ -> M.size midMap
+
+    log $ "size: big map (" <> show (M.size bigMap) <> ")"
+    benchWith 10  \_ -> M.size bigMap