addresses #100
adds benchmarks, addressing #101

Before:

```
size: singleton map
mean   = 7.85 μs
stddev = 62.54 μs
min    = 1.91 μs
max    = 1.60 ms
size: small map (100)
mean   = 303.09 μs
stddev = 482.41 μs
min    = 205.12 μs
max    = 3.85 ms
size: midsize map (10000)
mean   = 40.37 ms
stddev = 1.42 ms
min    = 36.60 ms
max    = 46.82 ms
size: big map (1000000)
mean   = 5.79 s
stddev = 60.76 ms
min    = 5.67 s
max    = 5.89 s
```

After:

```
size: singleton map
mean   = 3.00 μs
stddev = 38.96 μs
min    = 326.00 ns
max    = 1.16 ms
size: small map (100)
mean   = 10.16 μs
stddev = 7.21 μs
min    = 8.35 μs
max    = 175.09 μs
size: midsize map (10000)
mean   = 901.31 μs
stddev = 198.26 μs
min    = 822.48 μs
max    = 1.88 ms
size: big map (1000000)
mean   = 96.70 ms
stddev = 2.00 ms
min    = 92.99 ms
max    = 99.69 ms
```