@paf31 2017-06-05
Looks great, and thanks for adding the benchmarks!

@garyb What do you think about moving `minibench` into core?