@garyb 2017-06-05
> What do you think about moving `minibench` into core?

:100: :+1: