@paf31 2017-06-05
I just realized this is actually `O(n)`, which I documented in the release notes 😄 

Thanks again!