@matthewleon 2017-06-05
Good point!
On Mon, 5 Jun 2017 at 21:59, Phil Freeman <notifications@github.com> wrote:

> I just realized this is actually O(n), which I documented in the release
> notes 😄
>
> Thanks again!
>
> —
> You are receiving this because you authored the thread.
> Reply to this email directly, view it on GitHub
> <https://github.com/purescript/purescript-maps/pull/107#issuecomment-306308359>,
> or mute the thread
> <https://github.com/notifications/unsubscribe-auth/AARBSynTOX-tpvpQuaJHbS-Vy_pEEH0Tks5sBGwrgaJpZM4Nv61l>
> .
>